requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
babel==2.17.0
bitsandbytes==0.49.2
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
distro==1.9.0
clip @ git+https://github.com/ultralytics/CLIP.git@88ade288431a46233f1556d1e141901b3ef0a36b
comm==0.2.3
cuda-bindings==12.9.4
cuda-pathfinder==1.4.2
debugpy==1.8.17
defusedxml==0.7.1
ecdsa==0.19.1
einops==0.8.2
exceptiongroup==1.3.1
fairscale==0.4.13
fastapi==0.121.3
fastjsonschema==2.21.2
filelock==3.25.2
fire==0.7.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.2.0
ftfy==6.3.1
gdown==5.2.0
h11==0.16.0
hf-xet==1.4.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
ipykernel==7.1.0
ipywidgets==8.1.8
isoduration==20.11.0
Jinja2==3.1.6
jiter==0.12.0
json5==0.12.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
lap==0.5.12
lark==1.3.1
llama_stack==0.2.11
llama_stack_client==0.2.12
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mistune==3.1.4
mpmath==1.3.0
multidict==6.7.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.5.0
notebook_shim==0.2.4
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai==2.26.0
opencv-python==4.12.0.88
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
pillow==12.1.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
py-cpuinfo==9.0.0
pyaml==25.7.0
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PySocks==1.7.1
pysrt==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-json-logger==4.0.0
python-multipart==0.0.20
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scipy==1.15.3
Send2Trash==1.8.3
setuptools==82.0.1
sentencepiece==0.2.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
terminado==0.18.1
tiktoken==0.11.0
timm==1.0.25
tinycss2==1.4.0
tokenizers==0.20.3
tomli==2.3.0
torch==2.10.0
torchvision==0.25.0
tornado==6.5.2
tqdm==4.67.3
traitlets==5.14.3
transformers==4.46.3
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.163
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.6.3
wcwidth==0.6.0
webcolors==25.10.0
websocket-client==1.9.0
widgetsnbextension==4.0.15
yarl==1.22.0
